Update to newest neo4rs
They've added some serialization so that you can deserialize straight
to a custom struct for a row, which does clean up the deserialization.

The main issue I found was that I wasn't able to use `row.get` despite
it looking like it should also work, but I was always getting some
errors with BoltType. I also got the same with Biome, Tag and similar
structs to that. Might be something to investigate later.

But right now .to works on the row.