Swap to using serde_json::Number
For some reason, the x was not coming through correctly in some cases
from neo4j or through the library. No idea why.

But when I swap to the coordinate_3d that is also saved, everything
works... mostly. Other than it seems like sometimes it comes
through as -1.0 instead of -1 but when I use a serde_json::Number,
then everything is fine.